Move md5 string extension to own file
